combine-env-files.dev.sh, compose.dev.yaml, .dockerignore create. INSTALLATION update.
Create .env.dev file automatically from the ones in the services. Create compose file for reverse-proxy and for the full app with its 4 services (3+db). Update installation with instructions until full composed app testing. Configure Nginx server and offer instructions to test it independently and as part of the full app.